Rework GlobalFiltersExample
This avoids the deprecated AttributeElements and cleans it up a bit.
Also more important components were inlined so that they are visible
in the Show code section.

JIRA: RAIL-4350
